attendance api for use from gradebook and perhaps other trackers. needs shared secret ATTENDANCE_API_KEY
